% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_optimal.R
\name{A_Optimal}
\alias{A_Optimal}
\title{A-optimal Design}
\usage{
A_Optimal(N, A)
}
\arguments{
\item{N}{the number of feasible points}

\item{A}{a list of the design models, computed from the Polynomial Order, computed at each Feasible Point}
}
\value{
returns the CVXR solution to the optimization problem
A_Optimal(N,A)
}
\description{
Computes an A-optimal Design using simplex constraints
}
