% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_optimal.R
\name{C_Optimal}
\alias{C_Optimal}
\title{c-optimal Design}
\usage{
C_Optimal(N, A, c)
}
\arguments{
\item{N}{the number of feasible points}

\item{A}{a list of the design models, computed from the Polynomial Order, computed at each Feasible Point}

\item{c}{the scaling value, can be a vector or matrix}
}
\value{
returns the CVXR solution to the optimization problem
C_Optimal(N,A,c)
}
\description{
Computes an c-optimal Design using simplex constraints
}
