% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_optimal.R
\name{D_Optimal}
\alias{D_Optimal}
\title{D-optimal Design}
\usage{
D_Optimal(N, A)
}
\arguments{
\item{N}{the number of feasible points}

\item{A}{a list of N matrices from the regression model computed from the Polynomial Order, computed at each Feasible Point}
}
\value{
returns the CVXR solution to the optimization problem
D_Optimal(N,A)
}
\description{
Computes a D-optimal Design
}
