% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Computes the design model}
\usage{
model(N, a, b, PolyOrder, FeasiblePoints)
}
\arguments{
\item{N}{the number of feasible points}

\item{a}{the lower bound}

\item{b}{the upper bound}

\item{PolyOrder}{the order of the design space}

\item{Method}{the type of optimal design to be used}

\item{FeasablePoints}{the points whose weights will be optimized}
}
\value{
returns A the design model
}
\description{
Computes the design model
}
\examples{
model(N = 21, a = -1, b = 1, PolyOrder = 3, FeasiblePoints = function(i,a = -1,b = 1) a + (b-a) * (i-1) / (N-1))
model(N = 21, a = -1, b = 1, PolyOrder = 3, FeasiblePoints =  -1 + (1 - (-1)) * (1:21-1) / (21-1))
model(N,a,b,PolyOrder,FeasiblePoints)
}
