\name{OptimalDesign-internal}
        \alias{od.A.AA}
        \alias{od.A.KL}
        \alias{od.A.RC}
        \alias{od.A.RCs}
		\alias{od.A.SOCP}
		\alias{od.A.IQP}
		\alias{od.D.AA}
        \alias{od.D.KL}
        \alias{od.D.RC}
        \alias{od.D.RCs}
		\alias{od.D.SOCP}
		\alias{od.D.IQP}
		\alias{F.IVtoA}
		
		
        \title{Internal OptimalDesign Functions}
        \description{
          Internal OptimalDesign functions
        }
        \details{
          These are not to be called by the user.
        }
        \keyword{ internal } 