% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{optimalDesigns}
\alias{optimalDesigns}
\title{Optimal Design}
\usage{
optimalDesigns(
  N,
  a,
  b,
  PolyOrder = 3,
  Method = "D",
  FeasiblePoints = NA,
  A = NA,
  c = NA,
  ...
)
}
\arguments{
\item{N}{the number of feasible points}

\item{a}{the lower bound (numeric)}

\item{b}{the upper bound (numeric, where b > a)}

\item{PolyOrder}{the order of the design space}

\item{Method}{the type of optimal design to be used ('A','C','D','E')}

\item{FeasiblePoints}{the points whose weights will be optimized, can be in the form of a vector of size N, or as a function for points 1 to N}

\item{A}{a list of the design models, computed from the Polynomial Order, computed at each Feasible Point}

\item{c}{a scaling parameter to the model}
}
\value{
returns the CVXR solution to the optimization problem
}
\description{
Computes an Optimal Design by minimizing the design with simplex constraints
}
\examples{
optimalDesigns(N = 21, a = -1, b = 1,PolyOrder = 3, Method = "D")[[1]]
This will return the 21 interior point optimal weights according to a D-Optimal Design
}
