% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_optimal.R
\name{E_Optimal}
\alias{E_Optimal}
\title{E-optimal Design}
\usage{
E_Optimal(N, A)
}
\arguments{
\item{N}{the number of feasible points}

\item{A}{a list of N matrices from the regression model computed from the Polynomial Order, computed at each Feasible Point}
}
\value{
returns the CVXR solution to the optimization problem
E_Optimal(N,A)
}
\description{
Computes an E-optimal Design
}
