\name{OptimalDesign-internal}
        \alias{od_A_AQUA}
        \alias{od_A_DEL}
        \alias{od_A_KL}
        \alias{od_A_KL_bin}
        \alias{od_A_MISOCP}
        \alias{od_A_MUL}
        \alias{od_A_RC}
        \alias{od_A_REX}
	\alias{od_A_VDM}
	\alias{od_c_LP}
	\alias{od_c_MISOCP}
        \alias{od_D_AQUA}
        \alias{od_D_DEL}
        \alias{od_D_KL}
        \alias{od_D_KL_bin}
        \alias{od_D_MISOCP}
        \alias{od_D_MUL}
        \alias{od_D_RC}
        \alias{od_D_REX}
	\alias{od_D_VDM}
	\alias{ginv}	
	\alias{verify}	
        \title{Internal OptimalDesign Functions}
        \description{
          Internal OptimalDesign functions
        }
        \details{
          These are not to be called by the user.
        }
        \keyword{ internal } 