\name{OptimalDesign-package}

\alias{OptimalDesign-package}

\alias{OptimalDesign}

\docType{package}

\title{
  OptimalDesign
}

\description{
  Procedures for computing D-, A-, and IV-optimal approximate and exact designs of experiments 
  on finite domains, for regression models with real-valued, uncorrelated observations. 
}

\author{ 
  Radoslav Harman, Lenka Filova
}

\keyword{
  Optimal Design, A-optimality, D-optimality, IV-optimality
}
