\name{OptimalDesign-package}

\alias{OptimalDesign-package}

\alias{OptimalDesign}

\docType{package}

\title{
  OptimalDesign
}

\description{
  Procedures for computing D-, A-, I-, and c-optimal approximate and exact designs of experiments on finite domains, for regression models with uncorrelated observations. 
}

\author{ 
  Radoslav Harman, Lenka Filova
}

\keyword{ Optimal Design }
\keyword{ A-optimality }
\keyword{ D-optimality }
\keyword{ I-optimality }
\keyword{ c-optimality }

