\name{OptimalDesign-package}
\alias{OptimalDesign-package}
\alias{OptimalDesign}
\docType{package}
\title{
Optimal experimental designs.
}
\description{
A suite of tools to evaluate and find optimal experimental designs.
}
\details{
\tabular{ll}{
Package: \tab OptimalDesign\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-12-19\cr
License: \tab BSD_3_clause \cr
}
}
\author{
Michael Lerch, John Sherrill
}
\keyword{ package, experimental design }
